Day 12 Part 2 Solution

+++ day12/data.h
@@ -29,6 +29,6 @@ namespace AoC
 
     struct OutputData
     {
-        u32 PossibleArrangementsCount{};
+        u64 PossibleArrangementsCount{};
     };
 }
